+++ libs/events/events.go
@@ -35,7 +35,8 @@ type EventSwitch interface {
 	Fireable
 
 	// Multiple callbacks can be registered for each listener.
-	AddListener(listenerID string, cb EventCallback) error
+	AddListener(listenerID string, cb EventCallback)
+	Subscribe(listenerID string) <-chan Event
 	// Removes all callbacks for listener.
 	RemoveListener(listenerID string)
 }
@@ -56,6 +57,10 @@ type eventSwitch struct {
 	listeners []listenCell
 }
 
+func NilEventSwitch() EventSwitch {
+	return (*eventSwitch)(nil)
+}
+
 func NewEventSwitch() EventSwitch {
 	evsw := &eventSwitch{
 		listeners: make([]listenCell, 0, 10),
@@ -70,11 +75,37 @@ func (evsw *eventSwitch) OnStart() error {
 
 func (evsw *eventSwitch) OnStop() {}
 
-func (evsw *eventSwitch) AddListener(listenerID string, cb EventCallback) error {
+func (evsw *eventSwitch) AddListener(listenerID string, cb EventCallback) {
 	evsw.mtx.Lock()
 	evsw.listeners = append(evsw.listeners, listenCell{listenerID, cb})
 	evsw.mtx.Unlock()
-	return nil
+}
+
+// Returns a synchronous event emitter.
+func (evsw *eventSwitch) Subscribe(listenerID string) <-chan Event {
+	ch := make(chan Event, 0) // synchronous
+	return evsw.SubscribeOn(listenerID, ch)
+}
+
+// Like Subscribe, but lets the caller construct a channel.  If the capacity of
+// the provided channel is 0, it will be called synchronously; otherwise, it
+// will drop when the capacity is reached and a select doesn't immediately
+// send.
+func (evsw *eventSwitch) SubscribeOn(listenerID string, ch chan Event) <-chan Event {
+	evsw.AddListener(listenerID, func(event Event) {
+		// NOTE: This callback must not block.
+		if cap(ch) == 0 {
+			ch <- event // synchronous
+		} else {
+			select {
+			case ch <- event:
+			default: // async
+				evsw.RemoveListener(listenerID) // TODO log
+				close(ch)
+			}
+		}
+	})
+	return ch
 }
 
 func (evsw *eventSwitch) RemoveListener(listenerID string) {
@@ -89,7 +120,12 @@ func (evsw *eventSwitch) RemoveListener(listenerID string) {
 	evsw.mtx.Unlock()
 }
 
+// FireEvent on a nil switch is a noop, but no other operations are allowed for
+// safety.
 func (evsw *eventSwitch) FireEvent(event Event) {
+	if evsw == nil {
+		return
+	}
 	evsw.mtx.RLock()
 	listeners := make([]listenCell, len(evsw.listeners))
 	copy(listeners, evsw.listeners)